+++ packages/auto-complete/src/AutoCompleteStateless/examples/AutoCompleteStatelessEvents.tsx
@@ -37,7 +37,7 @@ export default class AutoCompleteStatelessEventsExample extends React.Component<
           onChange={(e) => {
             this.setState({ value: e.target.value, dataSource: getStaticData(e.target.value) });
           }}
-          onSelect={(value) => this.setState({ value })}
+          onSelect={(item: DataSourceItem) => this.setState({ value: item.label })}
           isFullWidth={false}
           before={<PaddedSearchIcon color="shade.500" />}
           onKeyDown={() => console.log('keyDown')}