+++ packages/address-picker/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog - @lendi-ui/address-picker
 
+## [1.0.0] 2019-6-27 [HUB-66](https://creditandfinance.atlassian.net/browse/HUB-66)
+### Created
+- Release version of AddressPicker.
+
 ## [0.1.0] 2019-6-27 [HUB-66](https://creditandfinance.atlassian.net/browse/HUB-66)
 ### Created
 - Created AddressModal component in AddressPicker.